refactor: extract useViewLayout composable from App.vue
isStackLayout, lastChatViewMode, restoreChatViewForSession,
displayedCurrentSessionId, and the isStackLayout→activePane watch.
